--- src/class.h
@@ -1,3 +1,0 @@
-class UnixSocket;
-class ISubject;
-class IObserver;
