+++ src/i_observer.cpp
@@ -1,0 +1,9 @@
+#include "i_observer.h"
+
+IObserver::~IObserver() {
+}
+
+void
+IObserver::update(int num) {
+}
+
