+++ src/i_subject.h
@@ -1,0 +1,17 @@
+#ifndef INCLUDED_SERVER
+#include "server.h"
+#endif
+
+class ISubject
+{
+public:
+    ISubject();
+    virtual ~ISubject();
+    virtual void subscribe(Server* server);
+    virtual void unsubscribe(Server* server);
+    virtual void notify(int);
+
+private:
+    Server* server_;
+};
+
